[BI-1363] shared ontology: add styling
