+++ src/components/modals/GenericModal.vue
@@ -1,0 +1,47 @@
+<!--
+  - See the NOTICE file distributed with this work for additional information
+  - regarding copyright ownership.
+  -
+  - Licensed under the Apache License, Version 2.0 (the "License");
+  - you may not use this file except in compliance with the License.
+  - You may obtain a copy of the License at
+  -
+  -     http://www.apache.org/licenses/LICENSE-2.0
+  -
+  - Unless required by applicable law or agreed to in writing, software
+  - distributed under the License is distributed on an "AS IS" BASIS,
+  - WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  - See the License for the specific language governing permissions and
+  - limitations under the License.
+  -->
+
+<template>
+  <BaseModal
+      v-bind:active.sync="active"
+      v-on:deactivate="$emit('deactivate')"
+  >
+    <h3 class="is-5 title has-text-info">
+      {{msgTitle}}
+    </h3>
+    <slot></slot>
+  </BaseModal>
+</template>
+
+<script lang="ts">
+import { Component, Prop, Vue } from 'vue-property-decorator';
+import BaseModal from '@/components/modals/BaseModal.vue';
+import { AlertCircleIcon } from 'vue-feather-icons'
+
+@Component({
+  components: {BaseModal, AlertCircleIcon}
+})
+export default class GenericModal extends Vue {
+  @Prop()
+  active!: boolean;
+  @Prop()
+  bodyClass!: Object;
+  @Prop()
+  msgTitle!: string;
+}
+
+</script>
